refactor: вынес часть логики
